requirements.txt
pandas~=0.25.3
matplotlib~=3.1.0
joblib~=0.14.0
scikit-learn~=0.21.3
requests~=2.22.0
Django~=3.0.7
django-heroku
djangorestframework~=3.11.0
gunicorn